Implement authorization using Gates

+++ app/Http/Controllers/QuestionsController.php
@@ -64,7 +64,9 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
-        return view('questions.edit', compact('question'));
+        if(\Gate::allows('update-question', $question))
+            return view('questions.edit', compact('question'));
+        abort(403,"Access denied");
     }
 
     /**
@@ -76,8 +78,9 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if(\Gate::denies('update-question', $question))
+            abort(403,"Access denied");
         $question->update($request->only('title','body'));
-
         return redirect()->route('questions.index')->with('success','Your question has been updated');
     }
 
@@ -89,6 +92,8 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(\Gate::denies('delete-question', $question))
+            abort(403,"Access denied");
         $question->delete();
         return redirect()->route('questions.index')->with('success','Your question has been deleted');
     }
